Fix on_exchange_declareok callback
Newer pika versions have changed the order of the arguments in
`exchange_declare`. So now when trying to set up an exchange we pass in
the wrong parameters. To prevent this from happening again I've made
them kwargs (also changed order a little to match the method signature)

See method signature for `exchange_declare` here:
https://pika.readthedocs.io/en/stable/_modules/pika/channel.html

Already fixed this once for `queue_declare` as well here:
https://github.com/ByteInternet/amqpconsumer/pull/12